pyproject.toml
[project]
name = "av-01"
version = "0.1.0"
description = ""
authors = [
    {name = "av",email = "av.avramenko@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "ruff (>=0.9.2,<0.10.0)",
    "mypy (>=1.14.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
